requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-jsonapi==6.1.0
django-filter==24.3
Pillow==9.4.0
psycopg2-binary==2.9.5
mysqlclient==2.1.1 
gunicorn==20.1.0  
django-cors-headers==4.1.0 
django-environ==0.11.2
drf-yasg==1.21.7
stripe==10.8.0
